Se agrego la validación de imagenes al editar un receta.

+++ includes/editar_recetas.php
@@ -5,12 +5,16 @@ if ($_SESSION['rol'] != 1) {
     exit;
 }
 
+// Generar una alerta
+$alertvalue = 0;
 $id = $_GET['id'];
 require_once "../conexion.php";
+
 // Procesar el formulario cuando se envía
 if ($_SERVER["REQUEST_METHOD"] == "POST") {
+    
     // Obtener datos del formulario
-    $id = $_POST['txtId'];
+    $idreceta = $_POST['txtId'];
     $idsubcategoria = $_POST['txtidsubcategoria'];
     $nombre = $_POST['txtNombre'];
     $Observaciones = $_POST['txtObservaciones'];
@@ -21,36 +25,42 @@ if ($_SERVER["REQUEST_METHOD"] == "POST") {
     // Manejar la carga de la nueva imagen si se proporciona
     if ($_FILES['txtImagen']['error'] == UPLOAD_ERR_OK && !empty($_FILES['txtImagen']['name'])) {
 
-        // Obtener la información de la imagen desde la base de datos
-        $query_imagen = mysqli_query($conexion, "SELECT imagen FROM recetas WHERE id = '$id'");
-        // Genera el arreglo de la fila en la tabla platos
-        $data_imagen = mysqli_fetch_assoc($query_imagen);
-        // Agregamos otra capa al directorio raíz para que las imagenes se eliminen
-        // MODIFICAAMOS CUANDO SE CREE LA CARPETA EDITAR DENTRO DE INCLUDES
-        // $imagenPath = "../" . $data_imagen['imagen'];
-        $imagenPath = $data_imagen['imagen'];
-        // Utilizamos el método unlink para borrar la imagen de la carpeta de recetas
-        unlink($imagenPath);
+        // Obtener la extensión de la imagen
+        $extension = pathinfo($_FILES['txtImagen']['name'], PATHINFO_EXTENSION);
 
-        // Mover la nueva imagen
-        $imagen_temp = $_FILES['txtImagen']['tmp_name'];
-        $imagen = "../assets/img/recetas/" . $_FILES['txtImagen']['name'];
-        move_uploaded_file($imagen_temp, $imagen);
+        // Validar la extensión permitida
+        $extensionesPermitidas = array('jpg', 'jpeg', 'png');
+        if (!in_array(strtolower($extension), $extensionesPermitidas)) {
+            $alertvalue = 1;
+            header("Location: editar_recetas.php?id=$idreceta&alertvalue=$alertvalue");
+            exit;  // Agregado para finalizar la ejecución después de la redirección
+        } else {
+            // Obtener la información de la imagen desde la base de datos
+            $query_imagen = mysqli_query($conexion, "SELECT imagen FROM recetas WHERE id = '$id'");
+            $data_imagen = mysqli_fetch_assoc($query_imagen);
+            $imagenPath = $data_imagen['imagen'];
+            unlink($imagenPath);
+
+            // Mover la nueva imagen
+            $imagen_temp = $_FILES['txtImagen']['tmp_name'];
+            $imagen = "../assets/img/recetas/" . $_FILES['txtImagen']['name'];
+            move_uploaded_file($imagen_temp, $imagen);
+        }
     } else {
         $imagen = $imagenActual;
-    }
-
-    // Validar y sanitizar los datos según sea necesario
-    // Actualizar la base de datos con los valores obtenidos del formulario
-    $query = "UPDATE recetas SET nombre = '$nombre', idsubcategoria = '$idsubcategoria', observaciones = '$Observaciones', imagen = '$imagen', fecha = '$fecha', estado = '$estado' WHERE id = $id";
-
-    if (mysqli_query($conexion, $query)) {
-        // La actualización fue exitosa, puedes redirigir a una página de éxito o realizar otras acciones
-        header('Location: ../vistas/crear_recetas.php');
-        exit;
-    } else {
-        // Hubo un error en la actualización, puedes mostrar un mensaje de error o redirigir a una página de error
-        $alert = "Error al actualizar la categoría: " . mysqli_error($conexion);
+        // Validar y sanitizar los datos según sea necesario
+        // Actualizar la base de datos con los valores obtenidos del formulario
+        $query = "UPDATE recetas SET nombre = '$nombre', idsubcategoria = '$idsubcategoria', observaciones = '$Observaciones', imagen = '$imagen', fecha = '$fecha', estado = '$estado' WHERE id = $idreceta";
+        if (mysqli_query($conexion, $query)) {
+            // La actualización fue exitosa, puedes redirigir a una página de éxito o realizar otras acciones
+            header('Location: ../vistas/crear_recetas.php');
+            exit;
+        } else {
+            // Hubo un error en la actualización, puedes mostrar un mensaje de error o redirigir a una página de error
+            $alertvalue = 2;
+            header("Location: editar_recetas.php?id=$idreceta&alertvalue=$alertvalue");
+            exit;  // Agregado para finalizar la ejecución después de la redirección
+        }
     }
 }
 
@@ -67,7 +77,27 @@ include_once "../includes/header.php";
                 <h4 class="card-title">Editar recetas</h4>
             </div>
             <div class="card-body">
-                <?php echo isset($alert) ? $alert : ''; ?>
+                <?php
+                //echo isset($_GET['alert']) ? htmlspecialchars($_GET['alert']) : ' '; 
+                $alert = isset($_GET['alertvalue']) ? htmlspecialchars($_GET['alertvalue']) : 0;
+                if ($alert == 1) {
+                    echo '<div class="alert alert-warning alert-dismissible fade show" role="alert">
+                           Solo se permiten archivos con extensiones JPG, JPEG y PNG
+                           <button type="button" class="close" data-dismiss="alert" aria-label="Close">
+                              <span aria-hidden="true">&times;</span>
+                           </button>
+                           </div>';
+                } elseif ($alert == 2) {
+                    echo '<div class="alert alert-warning alert-dismissible fade show" role="alert">
+                         Error al actualizar
+                         <button type="button" class="close" data-dismiss="alert" aria-label="Close">
+                             <span aria-hidden="true">&times;</span>
+                         </button>
+                         </div>';
+                } else {
+                    echo null;
+                }
+                ?>
                 <form action="../includes/editar_recetas.php" method="POST" class="p-1" enctype="multipart/form-data">
                     <input type="hidden" name="txtId" value="<?php echo $data['id'] ?>">
                     <!-- Editar la subcategoría, hace una consulta para traer las subcategorias de la tabla -->
@@ -102,16 +132,19 @@ include_once "../includes/header.php";
                         <?php
                         $imagenPath = $data['imagen'];
                         if (file_exists($imagenPath)) {
-                            echo '<img src="' . $imagenPath . '" alt="Imagen de receta" height="250px" class="btn-block" id="previewImagen">';
+                            echo '<div class="d-flex justify-content-center align-items-center">';
+                            echo '<img src="' . $imagenPath . '" alt="Imagen de receta" class="img-fluid" id="previewImagen" width="100%">';
+                            echo '</div>';
                         } else {
-                            echo '<img src="../assets/img/vacio.jpg" alt="Imagen de receta" height="250px" class="btn-block" id="previewImagen">';
-                            echo 'Imagen no disponible';
+                            echo '<div class="d-flex justify-content-center align-items-center">';
+                            echo '<img src="../assets/img/vacio.jpg" alt="Imagen de receta" class="img-fluid" id="previewImagen" width="100%">';
+                            echo '</div>';
                         }
                         ?>
                         <!-- Se agrega este input oculto para obtener la imagen actual enlazada a la base de datos -->
                         <input type="hidden" name="txtImagenActual" id="txtImagenActual" value="<?php echo $data['imagen']; ?>">
                         <!-- Cambia de enlace al input para abrir la imagen -->
-                        <a href="#" class="btn btn-primary btn-block" onclick="abrirImagenAgregar()">Modificar</a>
+                        <a href="#" class="btn btn-primary btn-block mt-3" onclick="abrirImagenAgregar()">Modificar</a>
                         <!-- Se agrega un input de tipo archivo oculto con el atributo accept -->
                         <input type="file" class="form-control" name="txtImagen" id="txtImagenInput" style="display: none" accept=".jpg, .jpeg, .png" onchange="mostrarPreview()">
                     </div>
@@ -132,20 +165,6 @@ include_once "../includes/header.php";
                         <button type="submit" class="btn btn-primary"><i class="fas fa-save"></i>Editar</button>
                         <a href="../vistas/crear_recetas.php" class="btn btn-danger">Cancelar</a>
                     </div>
-                    <div class="">
-                        <div class="form-group">
-                            <label for="estado" class=" text-dark font-weight-bold">Estado</label>
-                            <select class="form-control" name="txtEstado" id="txtEstado">
-                                <option value="1">Activada</option>
-                                <option value="0">Desactivada</option>
-                            </select>
-                        </div>
-                        <div>
-
-                            <button type="submit" class="btn btn-primary"><i class="fas fa-save"></i>Editar</button>
-                            <a href="../vistas/crear_recetas.php" class="btn btn-danger">Cancelar</a>
-                        </div>
-
                 </form>
             </div>
         </div>
@@ -173,9 +192,7 @@ include_once "../includes/header.php";
             reader.readAsDataURL(input.files[0]);
         }
     }
-
+    
 </script>
-
-</div>
 
 <?php include_once "../includes/footer.php"; ?>
